requirements.txt
agate==1.6.1
agate-dbf==0.2.0
agate-excel==0.2.2
agate-sql==0.5.3
alabaster==0.7.11
amqp==2.3.2
asn1crypto==0.24.0
Babel==2.6.0
bcrypt==3.1.4
billiard==3.5.0.4
boto3==1.9.26
botocore==1.12.26
CacheControl==0.11.7
celery==4.2.1
cement==2.10.2
census==0.8.7
cffi==1.11.5
click==6.7
cryptography==2.3
csvkit==1.0.3
dbfread==2.0.7
defusedxml==0.5.0
DictObject==1.0.5
dj-database-url==0.5.0
Django==2.0.8
django-cors-headers==2.4.0
django-environ==0.4.5
django-split-settings==0.3.0
django-storages==1.6.6
django-uuslug==1.1.8
djangorestframework==3.8.2
docutils==0.14
elex==2.4.2
et-xmlfile==1.0.1
Fabric3==1.14.post1
flake8==3.5.0
future==0.16.0
geographiclib==1.49
geojson==2.4.0
geopy==1.16.0
idna==2.7
imagesize==1.0.0
isodate==0.6.0
jdcal==1.4
jellyfish==0.5.6
Jinja2==2.10
jmespath==0.9.3
kombu==4.2.1
leather==0.3.3
lockfile==0.12.2
luckydonald-utils==0.58
MarkupSafe==1.0
mccabe==0.6.1
oauthlib==2.1.0
openpyxl==2.5.5
packaging==17.1
paramiko==2.4.1
parsedatetime==2.4
Pillow==5.2.0
-e git+https://github.com/The-Politico/politico-civic.git@ad4741836c445770d0e8305e1dcf0dc9dae4f62f#egg=politico_civic
politico-civic-almanac==0.1.0
politico-civic-ap-loader==0.3.0
politico-civic-biography==0.1.0
politico-civic-demography==0.1.2
politico-civic-election==0.2.1
politico-civic-election-night==0.12.5
politico-civic-entity==0.2.1
politico-civic-geography==0.5.1
politico-civic-government==0.1.2
politico-civic-race-ratings==0.2.6
politico-civic-stump==0.1.0
politico-civic-vote==0.1.4
psycopg2==2.7.5
pyasn1==0.4.4
pycodestyle==2.3.1
pycountry==18.5.26
pycparser==2.18
pyflakes==1.6.0
Pygments==2.2.0
PyJWT==1.6.4
pymongo==3.3.0
PyNaCl==1.2.1
pyparsing==2.2.0
pyshp==1.2.12
python-dateutil==2.2
python-dotenv==0.9.1
python-slugify==1.2.5
python-twitter==3.4.2
python3-openid==3.1.0
pytimeparse==1.1.8
pytz==2018.5
redis==2.10.6
requests==2.11.1
requests-oauthlib==1.0.0
s3transfer==0.1.13
Shapely==1.6.4.post2
six==1.11.0
slacker==0.9.65
snowballstemmer==1.2.1
social-auth-app-django==2.1.0
social-auth-core==1.7.0
Sphinx==1.7.6
sphinx-rtd-theme==0.4.1
sphinxcontrib-django==0.4
sphinxcontrib-websupport==1.1.0
SQLAlchemy==1.2.10
stringcase==1.2.0
tqdm==4.24.0
ujson==1.35
Unidecode==1.0.22
urllib3==1.23
us==1.0.0
vine==1.1.4
whitenoise==3.3.1
xlrd==1.1.0
